+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.1] - 2025-05-08
+
+### Added
+- Enhance error handling in get_logs endpoint
+
+## [1.5.0] - 2025-05-07
+
+### Added
+- Feat: add AWS region name to config
+
+## [1.4.0] - 2025-05-06
+
+### Added
+- Feat: logs router 
+
 ## [1.3.0] - 2025-04-24
 
 ### Refactored